refactor ValidatingVisitor to use factory pattern
