+++ liquibase-standard/pom.xml
@@ -342,6 +342,7 @@
                         <param>liquibase.changeset.ChangeSetService</param>
                         <param>liquibase.parser.LiquibaseSqlParser</param>
                         <param>liquibase.database.LiquibaseTableNames</param>
+                        <param>liquibase.changelog.visitor.ValidatingVisitor</param>
                     </services>
                 </configuration>
                 <executions>
