+++ liquibase-standard/src/main/java/liquibase/changelog/visitor/ValidatingVisitorFactory.java
@@ -1,0 +1,19 @@
+package liquibase.changelog.visitor;
+
+import liquibase.plugin.AbstractPluginFactory;
+
+public class ValidatingVisitorFactory extends AbstractPluginFactory<ValidatingVisitor> {
+    @Override
+    protected Class<ValidatingVisitor> getPluginClass() {
+        return ValidatingVisitor.class;
+    }
+
+    @Override
+    protected int getPriority(ValidatingVisitor obj, Object... args) {
+        return obj.getPriority();
+    }
+
+    public ValidatingVisitor getValidatingVisitor() {
+        return getPlugin();
+    }
+}
