+++ liquibase-standard/src/main/java/liquibase/util/ValidatingVisitorUtil.java
@@ -10,6 +10,7 @@ import liquibase.change.core.CreateIndexChange;
 import liquibase.change.core.DropIndexChange;
 import liquibase.change.core.SQLFileChange;
 import liquibase.changelog.*;
+import liquibase.changelog.visitor.StandardValidatingVisitor;
 import liquibase.database.Database;
 import liquibase.exception.DatabaseException;
 import liquibase.exception.UnexpectedLiquibaseException;
@@ -20,7 +21,7 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 /**
- * Util class to offload methods that are used by {@link liquibase.changelog.visitor.ValidatingVisitor} class
+ * Util class to offload methods that are used by {@link StandardValidatingVisitor} class
  * and may make it more complex than it should be
  */
 public class ValidatingVisitorUtil {
